Remove deprecated GroupBy methods
